Move reuseable methods from EmilysList scraper to base class
