+++ lib/scraper/emilys_list.rb
@@ -1,10 +1,8 @@
-require 'nokogiri'
-require 'httparty'
-require 'cta_aggregator_client'
-require 'Indirizzo'
+require 'scraper/scraper_base'
 
 module Scraper
-  class EmilysList
+  class EmilysList < ScraperBase
+
     ORIGIN_SYSTEM = "Emily's List"
     ORIGIN_URL = "http://www.emilyslist.org/pages/entry/events"
     EVENT_ATTRS = [
@@ -106,8 +104,6 @@ module Scraper
       date_node = find_start_date_node(title_node)
       schedule_node = page.at('p:contains("Schedule")')
       event['start_date'] = parse_start_date(date_node, schedule_node)
-      event['free'] = payment_button_present?(page)
-
       if date_lumped_with_location_data?(date_node)
         location_data = parse_node_data(date_node)
         location_data.shift
@@ -201,43 +197,6 @@ module Scraper
       return false
     rescue TypeError
       return false
-    end
-    
-    def find_next_node(current_node)
-      el = current_node.next_element
-
-      if el.children
-        el_children_with_data = el.children.reject { |child| child.text.gsub("\r\n", '').empty? }
-        if el_children_with_data.any?
-          el
-        else
-          el.next_element
-        end
-      end
-    end
-
-    def parse_location(location_data)
-      return if location_data.empty?
-      city_state_zip = Indirizzo::Address.new(location_data.pop)
-      location = {
-        address_lines: location_data,
-        locality: city_state_zip.city.first.titlecase,
-        region: city_state_zip.state,
-        postal_code: city_state_zip.zip
-      }
-    end
-
-    def is_zipcode?(candidate)
-      candidate.length == 5 && candidate.scan(/[[:digit:]]/).any?
-    end
-
-    def parse_node_data(node_data)
-      node_data.children.map { |data| data.text.strip  }.reject(&:empty?)
-    end
-
-    def load_webpage(link)
-      raw_page = HTTParty.get(link)
-      Nokogiri::HTML(raw_page)
     end
 
   end
